requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain-google-genai==1.0.8
langchain==0.2.11
psycopg2-binary==2.9.7
requests==2.31.0